Create simlarity_score
The file restaurant_reviews.txt is given to you. 

The file contains names of reviewers and their reviews of different restaurants in the comma-separated format. The first string is the name of the reviewer followed by the name of a restaurant and its rating. This Python program computes a similarity score between any two reviewers using Euclidean distances.